More code to integrate with ROS
Fixed a load of errors stemming from the modification of the message passing protocols and general code structure. Not sure if everything works, but the code is no longer throwing around errors. In order to make everything work properly the data coming from the sonar experiment simulator needs to be processed into the form that the particles are using.