Methods for running sonar sim
Pretty much duplicates of the move method in particle, which is itself a copy from the previous version of sonar sim. Might be a good idea to have a method which will do this movement given a start and end location and an angle. Will have to pass noise parameters into the math module though.