Fixed stupid bug
Was forgetting to normalise the value of the angle in particles. Since the simulator uses lines to do things, a value has to be present so that the lines are drawn and math is done on the lines in a way that makes sense - i.e. 0 degrees is 'up' on the sonar. I don't know how this will affect the results when using real sonar data. Still have not applied processing to it, so I don't know how best to transfer data.